Establish client-server connection, store & send message data
Message & unique ID assigned and logged when user connects.

Server listens for "chat" event from the client. When the client sends a chat message, the message is passed to the server as an argument to the "chat" event.

Server broadcasts the message to all connected clients & with a "received-chat" event, which is handled on the client side by socket.on() -- when event is triggered new chat is logged to console.

The inputMsg state variable stores user's input message, and the sendMessage function emits the chat event to the server with the message data.

The handleSubmit function creates a new message object with the user's name, message text, and timestamp. The setMsgList function is then used to update the state of the message list with the new message object, and the sendMessage function is called to emit the chat event to the server with the new message data.

So far everything is just in the console-- it needs to be displayed on the screen for the user to see.